refactor(emul-11): updated Layout.tsx
