--- src/components/layout.tsx
@@ -1,36 +1,0 @@
-/* eslint-disable */
-// This file was converted to .tsx without actually implementing typescript
-// @TODO: update this file to tsx and enable eslint
-import React from 'react'
-
-import './base/base.css'
-
-import { Header } from './organisms/site/Header/Header'
-import Footer from './organisms/footer/footer'
-
-class Template extends React.Component {
-  render() {
-    const { location, children, heroImage } = this.props
-    let home
-
-    if (location.pathname === '/') {
-      home = true
-    }
-
-    let rootPath = `/`
-    if (typeof __PREFIX_PATHS__ !== `undefined` && __PREFIX_PATHS__) {
-      rootPath = __PATH_PREFIX__ + `/`
-    }
-
-    return (
-      <div>
-        <Header image={heroImage} isHome={home} />
-        {children}
-        <Footer />
-      </div>
-    )
-  }
-}
-
-export default Template
-/* eslint-enable */
