+++ src/pages/index.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 import { graphql } from 'gatsby'
 import Img from 'gatsby-image'
 
-import Layout from '../components/layout'
+import { Layout } from '../components/templates/Layout'
 import SEO from '../components/base/seo/seo'
 
 import Card from '../components/molecules/card/card'
