+++ src/templates/case-study.tsx
@@ -6,7 +6,7 @@ import { Link, graphql } from 'gatsby'
 import get from 'lodash/get'
 import Img from 'gatsby-image'
 
-import Layout from '../components/layout'
+import { Layout } from '../components/templates/Layout'
 import SEO from '../components/base/seo/seo'
 import Share from '../components/molecules/share/share'
