+++ src/templates/landing-page.tsx
@@ -6,7 +6,7 @@ import { graphql } from 'gatsby'
 import get from 'lodash/get'
 import Img from 'gatsby-image'
 
-import Layout from '../components/layout'
+import { Layout } from '../components/templates/Layout'
 import SEO from '../components/base/seo/seo'
 
 import { Cta } from '../components/molecules/Cta/Cta'
